Convolver - largely working
Focus of change has been to get the Impulse resampler and normaliser working. Confirm the FFT being used and get the Overlap Save Convolver working. This has all been done. Also amended foxWavReader so that it now uses a generic reader rather than a specific file reader, which enables it to read for example the output from a shell call to SoX